+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     now make use of whitelist, blacklist, tags, and context. fixes issue #2
   * Added updated `validation()` method to PolySchema. fixes issue #3
   * Added `load` kwarg to Fields. Declares the key to lookup.
+  * Added `Child` field type -- Allows the use of nested values
   * fixed issue with PolySchema not getting it's own poly mapping instance
   * fixed issue with `name` not being respected on some missing values
   * `post_*` schema callables are now called even if no